MSSQLTest

+++ anyline-simple-data-jdbc-dialect/anyline-simple-data-jdbc-mssql/src/test/java/org/anyline/simple/mssql/MSSQLTest.java
@@ -86,7 +86,7 @@ public class MSSQLTest {
             row.put("NAME", "N"+i);
             set.add(row);
         }
-        int qty = service.insert(table, set);
+        long qty = service.insert(table, set);
         log.warn(LogUtil.format("[批量插入][影响行数:{}][生成主键:{}]", 36), qty, set.getStrings("ID"));
         Assertions.assertEquals(qty , 9);
